rename systray.js to systemtray.js to keep consistency
