+++ src/ags.src.gresource.xml
@@ -33,7 +33,7 @@
     <file>service/mpris.js</file>
     <file>service/network.js</file>
     <file>service/notifications.js</file>
-    <file>service/systray.js</file>
+    <file>service/systemtray.js</file>
 
     <file>dbus/types.js</file>
     <file>dbus/org.freedesktop.DBus.xml</file>
