+++ src/service/systemtray.ts
